moved workspace util and dependencyresolver to orchestra
