+++ maestro-orchestra/src/main/java/maestro/orchestra/workspace/DependencyResolver.kt
@@ -1,6 +1,5 @@
-package maestro.cli.util
+package maestro.orchestra.workspace
 
-import maestro.orchestra.ApplyConfigurationCommand
 import maestro.orchestra.CompositeCommand
 import maestro.orchestra.MaestroCommand
 import maestro.orchestra.yaml.MaestroFlowParser
@@ -14,43 +13,43 @@ object DependencyResolver {
     fun discoverAllDependencies(flowFile: Path): List<Path> {
         val discoveredFiles = mutableSetOf<Path>()
         val filesToProcess = mutableListOf(normalizePath(flowFile))
-        
+
         while (filesToProcess.isNotEmpty()) {
             val currentFile = normalizePath(filesToProcess.removeFirst())
-            
+
             // Skip if we've already processed this file (prevents circular references)
             if (discoveredFiles.contains(currentFile)) continue
-            
+
             // Add current file to discovered set
             discoveredFiles.add(currentFile)
-            
+
             try {
                 // Only process YAML files for dependency discovery
                 if (!isYamlFile(currentFile)) {
                     continue
                 }
-                
+
                 val flowContent = Files.readString(currentFile)
                 val commands = MaestroFlowParser.parseFlow(currentFile, flowContent)
-                
+
                 // Discover dependencies from each command
                 val dependencies = commands.flatMap { maestroCommand ->
                     extractDependenciesFromCommand(maestroCommand, currentFile)
                 }
-                
+
                 val newDependencies = dependencies
                     .map { normalizePath(it) }
                     .filter { it.exists() && !discoveredFiles.contains(it) }
                 filesToProcess.addAll(newDependencies)
-                
+
             } catch (e: Exception) {
                 // Ignore
             }
         }
-        
+
         return discoveredFiles.toList()
     }
-    
+
     private fun extractDependenciesFromCommand(maestroCommand: MaestroCommand, currentFile: Path): List<Path> {
         val commandDependencies = mutableListOf<Path>()
 
@@ -58,17 +57,17 @@ object DependencyResolver {
         maestroCommand.runFlowCommand?.let { runFlow ->
             resolveDependencyFile(currentFile, runFlow.sourceDescription)?.let { commandDependencies.add(it) }
         }
-        
+
         // Check for runScript commands and add the dependency if it exists (sourceDescription is not null)
         maestroCommand.runScriptCommand?.let { runScript ->
             resolveDependencyFile(currentFile, runScript.sourceDescription)?.let { commandDependencies.add(it) }
         }
-        
+
         // Check for retry commands and add the dependency if it exists (sourceDescription is not null)
         maestroCommand.retryCommand?.let { retry ->
             resolveDependencyFile(currentFile, retry.sourceDescription)?.let { commandDependencies.add(it) }
         }
-        
+
         // Check for assertScreenshot commands and add the reference image as a dependency
         maestroCommand.assertScreenshotCommand?.let { assertScreenshot ->
             resolveDependencyFile(currentFile, assertScreenshot.path)?.let { commandDependencies.add(it) }
@@ -80,7 +79,7 @@ object DependencyResolver {
                 resolveDependencyFile(currentFile, mediaPath)?.let { commandDependencies.add(it) }
             }
         }
-        
+
         // Handle configuration commands (onFlowStart, onFlowComplete)
         maestroCommand.applyConfigurationCommand?.let { config ->
             config.config.onFlowStart?.commands?.forEach { startCommand ->
@@ -90,7 +89,7 @@ object DependencyResolver {
                 commandDependencies.addAll(extractDependenciesFromCommand(completeCommand, currentFile))
             }
         }
-        
+
         // Handle ALL composite commands to extract dependencies from nested commands (RunFlow, Repeat, Retry)
         val command = maestroCommand.asCommand()
         if (command is CompositeCommand) {
@@ -98,20 +97,20 @@ object DependencyResolver {
                 commandDependencies.addAll(extractDependenciesFromCommand(nestedCommand, currentFile))
             }
         }
-        
+
         return commandDependencies
     }
-    
+
     private fun resolvePath(flowPath: Path, requestedPath: String): Path {
         val path = flowPath.fileSystem.getPath(requestedPath)
-        
+
         return if (path.isAbsolute) {
             path
         } else {
             flowPath.resolveSibling(path).toAbsolutePath()
         }
     }
-    
+
     private fun resolveDependencyFile(currentFile: Path, requestedPath: String?): Path? {
         val trimmed = requestedPath?.trim()
         if (trimmed.isNullOrEmpty()) return null
@@ -135,14 +134,14 @@ object DependencyResolver {
         val subflows = dependencies.filter { it != flowFile && isYamlFile(it) }
         val scripts = dependencies.filter { it != flowFile && isJsFile(it) }
         val otherFiles = dependencies.filter { it != flowFile && !isYamlFile(it) && !isJsFile(it) }
-        
+
         return buildString {
             appendLine("Dependency discovery for: ${flowFile.fileName}")
             appendLine("Total files: ${dependencies.size}")
             if (subflows.isNotEmpty()) appendLine("Subflows: ${subflows.size}")
             if (scripts.isNotEmpty()) appendLine("Scripts: ${scripts.size}")
             if (otherFiles.isNotEmpty()) appendLine("Other files: ${otherFiles.size}")
-            
+
             if (subflows.isNotEmpty()) {
                 appendLine("Subflow files:")
                 subflows.forEach { appendLine("  - ${it.fileName}") }
@@ -160,10 +159,8 @@ object DependencyResolver {
 
     private fun normalizePath(path: Path): Path {
         return try {
-            // Prefer canonical path without following symlinks
             path.toRealPath(LinkOption.NOFOLLOW_LINKS)
         } catch (e: Exception) {
-            // Fall back to absolute normalized path if real path resolution fails
             path.toAbsolutePath().normalize()
         }
     }
